wechat-miniprogram-workflow/requirements.txt
requests>=2.31.0
pydantic>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
structlog>=23.0.0
typing-extensions>=4.0.0
